nit: resolve PR comments
